reconnnected updated 0125

+++ src/main/java/com/qingcloud/iot/common/AppSdkMsgEvent.java
@@ -12,11 +12,11 @@ public class AppSdkMsgEvent {
         this.identifier = identifier;
     }
 
-    public Integer getTimestamp() {
+    public Long getTimestamp() {
         return timestamp;
     }
 
-    public void setTimestamp(Integer timestamp) {
+    public void setTimestamp(Long timestamp) {
         this.timestamp = timestamp;
     }
 
@@ -29,6 +29,6 @@ public class AppSdkMsgEvent {
     }
 
     public String identifier;
-    public Integer timestamp;
+    public Long timestamp;
     public HashMap<String,Object> params;
 }
